fix Cast one of the operands to a long bug
both of operands are int type which
can cause loss of data. This changes the types
of operands to long

+++ srv6/srv6-impl/src/main/java/io/fd/hc2vpp/srv6/read/policy/request/PolicyReadRequest.java
@@ -251,9 +251,9 @@ public class PolicyReadRequest extends JVppRequest
                         }));
     }
 
-    private Segment parseSrv6Sid(final int i, final Srv6Sid srv6Sid) {
+    private Segment parseSrv6Sid(final long i, final Srv6Sid srv6Sid) {
         // shifting index by 1 so it matches original indexing
-        long index = i + 1;
+        long index = i + 1L;
         SegmentBuilder builder = new SegmentBuilder().setKey(new SegmentKey(index)).setState(
                 new org.opendaylight.yang.gen.v1.http.cisco.com.ns.yang.oc.srte.policy.rev170918.segment.properties.StateBuilder()
                         .setIndex(index)
